+++ R/imports.R
@@ -3,10 +3,11 @@
 #' @import ggplot2
 #' @import ggthemes
 #' @import stringi
+#' @import tidyselect
 #' @importFrom magrittr "%>%"
-#' @importFrom rgeos gArea gIntersects gSimplify gIntersection gIsValid gBuffer
+#' @importFrom rgeos gArea gIntersects gSimplify gIntersection gIsValid gBuffer gContains
 #' @importFrom plotly ggplotly layout plotlyOutput event_data renderPlotly style
-#' @importFrom sp SpatialPolygonsDataFrame Polygons
+#' @importFrom sp SpatialPolygonsDataFrame Polygons SpatialPoints proj4string
 #' @importFrom tidyr gather
 #' @importFrom graphics text segments points polygon plot
 #' @importFrom stats dist
@@ -14,4 +15,5 @@
 #' @importFrom colourpicker colourInput
 #' @importFrom rgdal ogrListLayers readOGR
 #' @importFrom methods as
+#' @importFrom tibble column_to_rownames
 NULL